chg: refactor code
