--- src/libs/hasEllipsis/index.native.js
@@ -1,1 +1,0 @@
-export default () => {};
